refactor: 프론트엔드 수정
