+++ Backend/src/main/kotlin/com/luckyseven/backend/BackendApplication.kt
@@ -2,7 +2,11 @@ package com.luckyseven.backend
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
+import org.springframework.cache.annotation.EnableCaching
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing
 
+@EnableJpaAuditing
+@EnableCaching
 @SpringBootApplication
 class BackendApplication
