+++ Frontend/luckeyseven/src/pages/BudgetPage/BudgetPage.jsx
@@ -1,14 +1,13 @@
-import React, { useEffect, useState } from "react";
+import React, {useEffect, useState} from "react";
 import axios from 'axios';
-import { useParams } from "react-router-dom";
 import SetBudgetDialog from "./components/set-budget-dialog";
 import AddBudgetDialog from "./components/add-budget-dialog";
 import EditBudgetDialog from "./components/edit-budget-dialog";
 import PageHeaderControls from "../../components/PageHeaderControls";
-import { setBudgetInitialized } from "../../service/ApiService";
-import { currentTeamIdState } from "../../recoil/atoms/teamAtoms";
-import { useRecoilValue } from "recoil";
-import { SafeFormatterUtil } from './utils/SafeFormatterUtil';
+import {setBudgetInitialized} from "../../service/ApiService";
+import {currentTeamIdState} from "../../recoil/atoms/teamAtoms";
+import {useRecoilValue} from "recoil";
+import {SafeFormatterUtil} from './utils/SafeFormatterUtil';
 
 export function BudgetPage() {
   const teamId = useRecoilValue(currentTeamIdState);
@@ -22,7 +21,7 @@ export function BudgetPage() {
 
   const handleUpdate = async (updatePayload) => {
     try {
-      const res = await fetch(`/api/teams/${teamId}/budget`, {
+      const res = await fetch(`/api/teams/${teamId}/budgets`, {
         method: "PATCH",
         headers: {
           "Content-Type": "application/json",
@@ -71,7 +70,7 @@ export function BudgetPage() {
     try {
       setLoading(true);
 
-      const res = await axios.get(`/api/teams/${teamId}/budget`, {
+      const res = await axios.get(`/api/teams/${teamId}/budgets`, {
         headers: {
           "Authorization": `Bearer ${token}`,
         }
